Move matchers and helpers to their right place
* Helpers are modules mixed into the Cucumber world and are a preferred unit of reuse
  when you might otherwise have to refer to another step that isn't used in a feature.
* Matchers are just RSpec matchers
