+++ features/support/matchers/have_svg_text.rb
